backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic==2.9.2
email-validator==2.2.0
mangum==0.17.0
